Fixed a compile bug on platforms where a pointer does not fit in an int.
git-svn-id: http://yate.null.ro/svn/yate/trunk@3761 acf43c95-373e-0410-b603-e72c3f656dc1

+++ yateclass.h
@@ -4355,7 +4355,7 @@ public:
      * @return Valid array index
      */
     inline unsigned int index(void* ptr) const
-	{ return ((unsigned int)ptr) % m_length; }
+	{ return ((unsigned int)(unsigned long)ptr) % m_length; }
 
     /**
      * Retrieve the mutex associated with a given pointer.
